+++ libs/cx/src/lib/side-navigation/side-navigation.module.ts
@@ -7,9 +7,10 @@ import { SideNavigationUtilityDirective } from './side-navigation-utility.direct
 import { CxNestedListModule } from '@fundamental-ngx/cx/nested-list';
 import { SideNavigationButtonDirective } from './side-navigation-button.directive';
 import { ButtonModule } from '@fundamental-ngx/core/button';
+import { ScrollbarModule } from '@fundamental-ngx/core/scrollbar';
 
 @NgModule({
-    imports: [CommonModule, CxNestedListModule, ButtonModule],
+    imports: [CommonModule, CxNestedListModule, ButtonModule, ScrollbarModule],
     exports: [
         SideNavigationComponent,
         SideNavigationMainComponent,