+++ libs/docs/platform/settings-generator/examples/default/settings-generator-default-example.component.ts
@@ -51,7 +51,7 @@ class ExampleUserService {
             },
             {
                 value: 'de',
-                label: 'Deutch'
+                label: 'Deutsch'
             }
         ];
         return of(languages);