+++ libs/platform/src/lib/table/table.component.ts
@@ -48,7 +48,6 @@ import {
     ContentDensityObserver,
     contentDensityObserverProviders
 } from '@fundamental-ngx/core/content-density';
-import { PopoverComponent } from '@fundamental-ngx/core/popover';
 import { TableComponent as FdTableComponent, TableRowDirective } from '@fundamental-ngx/core/table';
 import { FDP_PRESET_MANAGED_COMPONENT, isDataSource, isString } from '@fundamental-ngx/platform/shared';
 import { cloneDeep, get } from 'lodash-es';
@@ -214,11 +213,13 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
     @Input()
     enableDragResize = true;
 
-    /** Whether to fix the table header and footer. Default is false.
+    /**
+     * Whether to fix the table header and footer. Default is true.
      * Note that if the table contains freezable columns, the header and
-     * footer will be fixed automatically. */
+     * footer will be fixed automatically, regardless of this input value.
+     * */
     @Input()
-    fixed = false;
+    fixed = true;
 
     /**
      * Table data source.
@@ -332,6 +333,10 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
 
     /** @Hidden */
     private _noBorderY = false;
+
+    /** Whether to allow for row reordering on tree tables via drag and drop. */
+    @Input()
+    enableRowReordering = true;
 
     /** Initial visible columns. Consist of a list of unique column names */
     @Input()
@@ -546,10 +551,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
     /** @hidden */
     @ViewChild('tableScrollable')
     readonly tableScrollable: TableScrollable;
-
-    /** @hidden */
-    @ViewChildren('columnHeaderPopover')
-    readonly columnHeaderPopovers: QueryList<PopoverComponent>;
 
     /** @hidden */
     @ViewChild('tableContainer')
@@ -1149,9 +1150,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
 
         this._tableService.freezeTo(columnName, end);
         this.recalculateTableColumnWidth();
-        this.columnHeaderPopovers.forEach((popover) => {
-            popover.close();
-        });
     }
 
     /** Unfreeze column */
@@ -1166,9 +1164,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
 
         this._tableService.freezeTo(freezeToPreviousColumnName, end);
         this.recalculateTableColumnWidth();
-        this.columnHeaderPopovers.forEach((popover) => {
-            popover.close();
-        });
     }
 
     /** expand all rows */
@@ -1438,6 +1433,58 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
     // Private API
 
     /** @hidden */
+    _scrollToOverlappedCell(): void {
+        const tableScrollableEl = this.tableScrollable.getElementRef().nativeElement;
+
+        if (
+            (this._freezableColumns.size || this._freezableEndColumns.size) &&
+            tableScrollableEl.scrollWidth > tableScrollableEl.clientWidth
+        ) {
+            const activeEl = document.activeElement;
+            if (
+                activeEl &&
+                !(
+                    activeEl.classList.contains('fd-table__cell--fixed') ||
+                    activeEl.classList.contains('fd-table__cell--fixed-end')
+                )
+            ) {
+                if (this._freezableColumns.size && !this._freezableEndColumns.size) {
+                    activeEl.scrollIntoView({ block: 'nearest', inline: 'end' });
+                } else if (!this._freezableColumns.size && this._freezableEndColumns.size) {
+                    activeEl.scrollIntoView({ block: 'nearest', inline: 'center' });
+                } else if (this._freezableColumns.size && this._freezableEndColumns.size) {
+                    // check to see if the active element is obstructed by another element
+                    const activeElLeft = activeEl.getBoundingClientRect().left;
+                    const activeElTop = activeEl.getBoundingClientRect().top;
+                    const topElementFromLeft = document.elementFromPoint(activeElLeft, activeElTop);
+                    // if the activeEl is overlapped
+                    if (
+                        topElementFromLeft &&
+                        !activeEl.isSameNode(topElementFromLeft) &&
+                        topElementFromLeft.classList.contains('fd-table__cell--fixed-end')
+                    ) {
+                        const topElementX = topElementFromLeft.getBoundingClientRect().left;
+                        const leftVal = this._rtl
+                            ? (activeElLeft + activeEl.getBoundingClientRect().width - topElementX) * -1
+                            : activeElLeft + activeEl.getBoundingClientRect().width - topElementX;
+                        tableScrollableEl.scrollBy({ top: 0, left: leftVal });
+                    } else if (
+                        topElementFromLeft &&
+                        !activeEl.isSameNode(topElementFromLeft) &&
+                        topElementFromLeft.classList.contains('fd-table__cell--fixed')
+                    ) {
+                        const topElementX = topElementFromLeft.getBoundingClientRect().right;
+                        const leftVal = this._rtl
+                            ? (activeElLeft - activeEl.getBoundingClientRect().width - topElementX) * -1
+                            : activeElLeft - activeEl.getBoundingClientRect().width - topElementX;
+                        tableScrollableEl.scrollBy({ top: 0, left: leftVal });
+                    }
+                }
+            }
+        }
+    }
+
+    /** @hidden */
     _isColumnHasHeaderMenu(column: TableColumn): boolean {
         return (
             column.sortable ||
@@ -1568,10 +1615,20 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
      * @hidden
      * Create table rows rearrange event
      */
-    _emitRowsRearrangeEvent(row: TableRow, previousIndex: number, newIndex: number): void {
+    _emitRowsRearrangeEvent(row: TableRow, dropRow: TableRow, event: FdDropEvent<TableRow>): void {
         const rows = this._tableRows.map(({ value }) => value);
 
-        this.rowsRearrange.emit(new TableRowsRearrangeEvent(row.value, previousIndex, newIndex, rows));
+        this.rowsRearrange.emit(
+            new TableRowsRearrangeEvent(
+                row.value,
+                dropRow.value,
+                event.draggedItemIndex,
+                event.replacedItemIndex,
+                event.insertAt,
+                event.mode,
+                rows
+            )
+        );
     }
 
     /**
@@ -1585,7 +1642,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
         } else {
             this.group([{ field, direction: SortDirection.NONE, showAsColumn: true }]);
         }
-        this._closePopoverForColumnByFieldName(field);
     }
 
     /**
@@ -1606,7 +1662,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
         } else {
             this.removeFilter([field]);
         }
-        this._closePopoverForColumnByFieldName(field);
     }
 
     /**
@@ -1615,7 +1670,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
      */
     _columnHeaderSortBy(field: string, direction: SortDirection): void {
         this.sort([{ field, direction }]);
-        this._closePopoverForColumnByFieldName(field);
     }
 
     /** @hidden */
@@ -1677,9 +1731,15 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
         );
     }
 
-    /** @hidden */
-    _isTreeRowFirstCell(cellIndex: number, row: TableRow): boolean {
-        return cellIndex === 0 && this._isTreeRow(row);
+    /**
+     * This method is used in both css class condition and in the (keydown.enter) handler. We want
+     * to make sure that we also allow keyboard handling for the non-tree cells.
+     *
+     * @param event Optional event that is used only for the keydown even handler
+     * @hidden
+     */
+    _isTreeRowFirstCell(cellIndex: number, row: TableRow, event?: Event): boolean {
+        return (cellIndex === 0 && this._isTreeRow(row)) || (!!event && !this._isTreeRow(row));
     }
 
     /** @hidden */
@@ -1713,7 +1773,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
             }
 
             this._cdr.markForCheck();
-            this._emitRowsRearrangeEvent(dragRow, event.draggedItemIndex, event.replacedItemIndex);
+            this._emitRowsRearrangeEvent(dragRow, dropRow, event);
         }
     }
 
@@ -1744,8 +1804,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
         nestingLevel: number | null
     ): Promise<void> {
         this._focusedCellPosition = { rowIndex: position.rowIndex, colIndex: position.colIndex };
-
-        this._scrollToOverlappedCell();
 
         if (this.cellFocusedEventAnnouncer) {
             this._liveAnnouncer.clear();
@@ -1779,14 +1837,6 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
             setTimeout(() => {
                 (event.target as HTMLElement).focus();
             });
-        }
-    }
-
-    /** @hidden */
-    private _closePopoverForColumnByFieldName(field: string): void {
-        const index = this._visibleColumns.findIndex((c) => c.key === field);
-        if (index !== -1) {
-            this.columnHeaderPopovers.get(index)?.close();
         }
     }
 
@@ -1987,7 +2037,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
         this._subscriptions.add(
             this._tableService.freezeChange.subscribe((event: FreezeChange) => {
                 this.columnFreeze.emit(new TableColumnFreezeEvent(this, event.current, event.previous));
-                this.fixed = !!this._freezableColumns.size || !!this._freezableEndColumns.size;
+                this.fixed = !!this.fixed || !!this._freezableColumns.size || !!this._freezableEndColumns.size;
             })
         );
 
@@ -2268,7 +2318,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
     private _setFreezableInfo(): void {
         this._freezableColumns = this._getFreezableColumns();
         this._freezableEndColumns = this._getFreezableEndColumns();
-        this.fixed = !!this._freezableColumns.size || !!this._freezableEndColumns.size;
+        this.fixed = !!this.fixed || !!this._freezableColumns.size || !!this._freezableEndColumns.size;
     }
 
     /** @hidden */
@@ -2764,6 +2814,7 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
                     this.recalculateTableColumnWidth();
                     if (this._freezableColumns.size || this._freezableEndColumns.size) {
                         this._tableColumnResizeService.updateFrozenColumnsWidth();
+                        this._cdr.detectChanges();
                     }
                 })
         );
@@ -2869,57 +2920,5 @@ export class TableComponent<T = any> extends Table<T> implements AfterViewInit, 
             row: ${position.rowIndex + 1} of ${position.totalRows}` +
             (nestingLevel !== null ? `, level: ${nestingLevel + 1}` : '')
         );
-    }
-
-    /** @hidden */
-    private _scrollToOverlappedCell(): void {
-        const tableScrollableEl = this.tableScrollable.getElementRef().nativeElement;
-
-        if (
-            (this._freezableColumns.size || this._freezableEndColumns.size) &&
-            tableScrollableEl.scrollWidth > tableScrollableEl.clientWidth
-        ) {
-            const activeEl = document.activeElement;
-            if (
-                activeEl &&
-                !(
-                    activeEl.classList.contains('fd-table__cell--fixed') ||
-                    activeEl.classList.contains('fd-table__cell--fixed-end')
-                )
-            ) {
-                if (this._freezableColumns.size && !this._freezableEndColumns.size) {
-                    activeEl.scrollIntoView({ block: 'nearest', inline: 'end' });
-                } else if (!this._freezableColumns.size && this._freezableEndColumns.size) {
-                    activeEl.scrollIntoView({ block: 'nearest', inline: 'center' });
-                } else if (this._freezableColumns.size && this._freezableEndColumns.size) {
-                    // check to see if the active element is obstructed by another element
-                    const activeElLeft = activeEl.getBoundingClientRect().left;
-                    const activeElTop = activeEl.getBoundingClientRect().top;
-                    const topElementFromLeft = document.elementFromPoint(activeElLeft, activeElTop);
-                    // if the activeEl is overlapped
-                    if (
-                        topElementFromLeft &&
-                        !activeEl.isSameNode(topElementFromLeft) &&
-                        topElementFromLeft.classList.contains('fd-table__cell--fixed-end')
-                    ) {
-                        const topElementX = topElementFromLeft.getBoundingClientRect().left;
-                        const leftVal = this._rtl
-                            ? (activeElLeft + activeEl.getBoundingClientRect().width - topElementX) * -1
-                            : activeElLeft + activeEl.getBoundingClientRect().width - topElementX;
-                        tableScrollableEl.scrollBy({ top: 0, left: leftVal });
-                    } else if (
-                        topElementFromLeft &&
-                        !activeEl.isSameNode(topElementFromLeft) &&
-                        topElementFromLeft.classList.contains('fd-table__cell--fixed')
-                    ) {
-                        const topElementX = topElementFromLeft.getBoundingClientRect().right;
-                        const leftVal = this._rtl
-                            ? (activeElLeft - activeEl.getBoundingClientRect().width - topElementX) * -1
-                            : activeElLeft - activeEl.getBoundingClientRect().width - topElementX;
-                        tableScrollableEl.scrollBy({ top: 0, left: leftVal });
-                    }
-                }
-            }
-        }
     }
 }